requirements.txt
PyMuPDF
python-docx
Pillow~=10.3.0
pytesseract~=0.3.10
transformers~=4.41.2
sentence-transformers
streamlit~=1.36.0
numpy~=1.26.4

psycopg2~=2.9.9
openpyxl~=3.1.4
requests~=2.32.3
openai~=1.35.3
torch~=2.3.1
scikit-learn~=1.5.0
docx2txt~=0.8
pdf2image~=1.17.0
docx2pdf~=0.1.8